Updated mongo backend to work with mongo > 1.8.3
Starting in Mongo 1.8.4 the keys of the auths hash are symbols. The qu-mongo backend probably shouldn't care how mongo keys its hashes so I updated the code to avoid manipulating auths directly.

Also, the Connection class appears to be deprecated so I updated this to use the replacement MongoClient instead.