Create separate class to execute jobs
This is a proposal for #5. Creates a separate
class to execute the job instead of doing the
execution inside the worker. This allows us to
have different execution mechanisms, from the
current "just run", to a forking runner and
in the future other threaded runners (that
would be better suited for JRuby and Rubinius).